backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
aiomysql==0.2.0
mysqlclient==2.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
flower==2.0.1
geoalchemy2==0.14.2
shapely==2.0.2
geojson==3.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
email-validator==2.1.0
Pillow==10.1.0
numpy==1.26.2